+++ server/pg_models/stripe_payment.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
     const StripePayment = sequelize.define('StripePayment', {
         id: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.BIGINT,
             allowNull: false,
             primaryKey: true,
             autoIncrement: true
@@ -14,16 +14,16 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING
         },
         invoiceId: {
-            type: DataTypes.INTEGER
+            type: DataTypes.BIGINT
         },
         customerId: {
             type: DataTypes.STRING
         },
         amount: {
-            type: DataTypes.INTEGER
+            type: DataTypes.BIGINT
         },
         status: {
-            type: DataTypes.INTEGER
+            type: DataTypes.BIGINT
         },
         email: {
             type: DataTypes.STRING
@@ -35,7 +35,7 @@ module.exports = (sequelize, DataTypes) => {
             type: DataTypes.STRING
         },
         stripeConnectId: {
-            type: DataTypes.INTEGER
+            type: DataTypes.BIGINT
         },
     });
